fixed issue with finding patj

+++ controllers/developer/deleteApps.js
@@ -12,8 +12,8 @@ exports.deleteAppsRoute = async (req, res) => {
         let deleteDate = oldPublishedDate.setDate(oldPublishedDate.getDate() + 7)
         let today =new Date();
 
-        // if(new Date(deleteDate).getTime() < today.getTime())
-        // {
+        if(new Date(deleteDate).getTime() < today.getTime())
+        {
             rimraf.sync(`uploads/${theApp[0].appName}`);
             const result = await Promise.resolve(Apps.deleteApps(req.params.id, res.locals.dev.id)).catch(err =>{console.log(err)});
             if(result[0].affectedRows)
@@ -24,13 +24,11 @@ exports.deleteAppsRoute = async (req, res) => {
             if(fs.existsSync(`updates/${theApp[0].appName}`))
             {
                 rimraf.sync(`updates/${theApp[0].appName}`);
-            }else{
-                res.render("developer/appls/deleteApps")
-             }
-        // }else{
-            // let oneWeeklimit = "not enghough date"
-            // res.render("developer/appls/deleteApps", {oneWeeklimit:oneWeeklimit})
-        // }
+            }
+        }else{
+            let oneWeeklimit = "not enghough date"
+            res.render("developer/appls/deleteApps", {oneWeeklimit:oneWeeklimit})
+        }
 
     })
